wgpu-core: Move device/mod.rs's Global impl to device/global.rs.
